requirements/base.txt
django>=3.2
django-allauth
django-admin-inline-paginator==0.1
djangorestframework==3.12.2
django-rest-auth==0.9.1
django-rest-knox==4.1.0
django-cors-headers==3.7.0
django-extensions==3.1.0
drf-yasg==1.20

redis==3.5.3
celery==5.1.2
django-celery-beat==2.2.1

pillow==8.1.1
psycopg2-binary==2.8.6

git+https://github.com/Rock-n-Block/ipfs_lib.git@develop

Markdown==3.3.6
MarkupSafe==2.1.0

ethereum==2.0.4
eth-account==0.4.0
eth-utils==1.10.0
eth-abi==2.1.1
eth-keys==0.2.4
eth-keyfile==0.5.1
eth-hash==0.3.1
eth-rlp==0.2.1
rlp==0.6.0
web3 @ git+https://github.com/Rock-n-Block/web3.py.git
trx-utils==1.0.4
tronapi @ git+https://github.com/MyWishPlatform/tron-api-python.git@21efb1b359c18c3ab17efcc2bb324dffada7ec9a
httpcore
httpx
ecdsa==0.17.0

ptpython
marshmallow_dataclass
marshmallow-enum
pyyaml
eth-keys==0.2.4
eth-keyfile==0.5.1
eth-hash==0.3.1
ecdsa==0.17.0
pytest-django==4.4.0
mixer==7.2.0
Faker==12.0.0
pyTelegramBotAPI==4.1.1

sentry-sdk==1.5.7

asyncio==3.4.3
aioredis==2.0.1
websockets==10.3
django-sesame==3.1
watchdog==2.1.9

django-quill-editor
beautifulsoup4==4.11.1
filetype
model_bakery

importlib-metadata>=1.4.0,<5.0; python_version < '3.8'
django-nested-inline==0.4.5